Fixed and improved error handling
Using atomic.Value for storing and checking an error was an incorrect approach.

Error values would be treated as a different type, so adding multiple errors if
they came in prior to the error check would have caused a panic.

Replaced with x/sync/errgroup, a far better way of handing goroutines and
errors.